chore: fix anvil tests
